test(frontend): share mock API fixtures between Playwright and Vitest
The Vitest harness's mockApiFetch.ts started as a port of the Playwright
suite's mockApi.ts, leaving the same fixtures and /api/v1 route matching
defined twice. The two copies could drift silently (route shapes, provider
identity fields, capability flags), making the suites disagree about the
API contract they mock. Flagged by roborev review job 1703 on 2c43481c.

src/test/mockApiFetch.ts is now the single source: fixtures plus a
transport-neutral createMockApiHandler, with the fetch adapter built on
top. tests/e2e/support/mockApi.ts becomes a thin page.route adapter over
the same handler (it keeps only the SSE /events fulfillment, which the
jsdom suite replaces with an EventSource stub). The shared module lives
under src/test/ because the lint-api-urls hook only permits raw /api/v1
strings in test directories, and it must not import @playwright/test
types. Routes the fetch port had skipped (/repos/summary, legacy issue
sync POST, legacy PR PATCH) were restored so the merged core is
route-for-route identical to the old Playwright mock; spec-level
page.route overrides still shadow the defaults since later registrations
win.

Validation: vp test run (2318 passed); playwright --list collects 332
tests in 19 files; repo-summary-results-label and theme-toggle specs run
green in both browsers through the new adapter (the former exercises the
newly shared /repos/summary route); typecheck errors identical to clean
HEAD (pre-existing, none in touched files).
